requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.10.1
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.11
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
passlib==1.7.4
propcache==0.3.0
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
PyPDF2==3.0.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.18.3
